feat: rename folder Insight to insight
